// supabase/functions/write-report/types.ts
export type Json = string | number | boolean | null | {[key: string]: Json | undefined} | Json[];

export type Database = {
  public: {
    Tables: {
      api_keys: {
        Row: {
          created_at: string | null;
          description: string | null;
          expires_at: string | null;
          id: string;
          key: string;
          last_used_at: string | null;
          project_id: string;
          status: Database['public']['Enums']['api_key_status'];
          tenant_id: string;
        };
        Insert: {
          created_at?: string | null;
          description?: string | null;
          expires_at?: string | null;
          id?: string;
          key: string;
          last_used_at?: string | null;
          project_id: string;
          status?: Database['public']['Enums']['api_key_status'];
          tenant_id: string;
        };
        Update: {
          created_at?: string | null;
          description?: string | null;
          expires_at?: string | null;
          id?: string;
          key?: string;
          last_used_at?: string | null;
          project_id?: string;
          status?: Database['public']['Enums']['api_key_status'];
          tenant_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'api_keys_project_id_fkey';
            columns: ['project_id'];
            isOneToOne: false;
            referencedRelation: 'projects';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'api_keys_tenant_id_fkey';
            columns: ['tenant_id'];
            isOneToOne: false;
            referencedRelation: 'tenants';
            referencedColumns: ['id'];
          },
        ];
      };
// ...
      reports: {
        Row: {
          branch: string | null;
          commit_hash: string | null;
          creator_id: string | null;
          creator_type: Database['public']['Enums']['creator_type_enum'] | null;
          duration: number | null;
          id: string;
          project_id: string;
          raw_json: Json;
          run_at: string | null;
          status: Database['public']['Enums']['run_status_enum'] | null;
          tenant_id: string;
          test_tool: Database['public']['Enums']['test_tool_enum'];
        };
        Insert: {
          branch?: string | null;
          commit_hash?: string | null;
          creator_id?: string | null;
          creator_type?: Database['public']['Enums']['creator_type_enum'] | null;
          duration?: number | null;
          id?: string;
          project_id: string;
          raw_json: Json;
          run_at?: string | null;
          status?: Database['public']['Enums']['run_status_enum'] | null;
          tenant_id: string;
          test_tool: Database['public']['Enums']['test_tool_enum'];
        };
        Update: {
          branch?: string | null;
          commit_hash?: string | null;
          creator_id?: string | null;
          creator_type?: Database['public']['Enums']['creator_type_enum'] | null;
          duration?: number | null;
          id?: string;
          project_id?: string;
          raw_json?: Json;
          run_at?: string | null;
          status?: Database['public']['Enums']['run_status_enum'] | null;
          tenant_id?: string;
          test_tool?: Database['public']['Enums']['test_tool_enum'];
        };
        Relationships: [
          {
            foreignKeyName: 'reports_project_id_fkey';
            columns: ['project_id'];
            isOneToOne: false;
            referencedRelation: 'projects';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'reports_tenant_id_fkey';
            columns: ['tenant_id'];
            isOneToOne: false;
            referencedRelation: 'tenants';
            referencedColumns: ['id'];
          },
        ];
      };
// ...
    };
// ...
  };
};

type DefaultSchema = Database[Extract<keyof Database, 'public'>];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema['Enums'] | {schema: keyof Database},
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database;
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions['schema']]['Enums']
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {schema: keyof Database}
  ? Database[DefaultSchemaEnumNameOrOptions['schema']]['Enums'][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema['Enums']
    ? DefaultSchema['Enums'][DefaultSchemaEnumNameOrOptions]
    : never;